Dumb InstallButton that breaks everything
