+++ tests/client/disco/components/TestInstallButton.js
@@ -1,0 +1,136 @@
+import React from 'react';
+import { Simulate, renderIntoDocument } from 'react-addons-test-utils';
+import { findDOMNode } from 'react-dom';
+
+import {
+  InstallButton,
+} from 'disco/components/InstallButton';
+import {
+  DOWNLOADING,
+  INSTALLED,
+  INSTALLING,
+  THEME_TYPE,
+  UNINSTALLED,
+  UNINSTALLING,
+  UNKNOWN,
+} from 'disco/constants';
+import { getFakeI18nInst } from 'tests/client/helpers';
+
+
+describe('<InstallButton />', () => {
+  function renderButton(props = {}) {
+    const renderProps = {
+      dispatch: sinon.spy(),
+      install: sinon.spy(),
+      installTheme: sinon.spy(),
+      uninstall: sinon.spy(),
+      i18n: getFakeI18nInst(),
+      ...props,
+    };
+
+    return renderIntoDocument(
+      <InstallButton {...renderProps} />);
+  }
+
+  it('should be disabled if isDisabled status is UNKNOWN', () => {
+    const button = renderButton({status: UNKNOWN});
+    const root = findDOMNode(button);
+    const checkbox = root.querySelector('input[type=checkbox]');
+    assert.equal(checkbox.hasAttribute('disabled'), true);
+    assert.ok(root.classList.contains('unknown'));
+  });
+
+  it('should reflect UNINSTALLED status', () => {
+    const button = renderButton({status: UNINSTALLED});
+    const root = findDOMNode(button);
+    const checkbox = root.querySelector('input[type=checkbox]');
+    assert.equal(checkbox.hasAttribute('disabled'), false);
+    assert.ok(root.classList.contains('uninstalled'));
+  });
+
+  it('should reflect INSTALLED status', () => {
+    const button = renderButton({status: INSTALLED});
+    const root = findDOMNode(button);
+    const checkbox = root.querySelector('input[type=checkbox]');
+    assert.equal(checkbox.checked, true, 'checked is true');
+    assert.ok(root.classList.contains('installed'));
+  });
+
+  it('should reflect download downloadProgress', () => {
+    const button = renderButton({status: DOWNLOADING, downloadProgress: 50});
+    const root = findDOMNode(button);
+    assert.ok(root.classList.contains('downloading'));
+    assert.equal(root.getAttribute('data-download-progress'), 50);
+  });
+
+  it('should reflect installation', () => {
+    const button = renderButton({status: INSTALLING});
+    const root = findDOMNode(button);
+    assert.ok(root.classList.contains('installing'));
+  });
+
+  it('should reflect uninstallation', () => {
+    const button = renderButton({status: UNINSTALLING});
+    const root = findDOMNode(button);
+    assert.ok(root.classList.contains('uninstalling'));
+  });
+
+  it('should not call anything on click when neither installed or uninstalled', () => {
+    const install = sinon.stub();
+    const uninstall = sinon.stub();
+    const button = renderButton({status: DOWNLOADING, install, uninstall});
+    const root = findDOMNode(button);
+    Simulate.click(root);
+    assert.ok(!install.called);
+    assert.ok(!uninstall.called);
+  });
+
+  it('should associate the label and input with id and for attributes', () => {
+    const button = renderButton({status: UNINSTALLED, slug: 'foo'});
+    const root = findDOMNode(button);
+    assert.equal(root.querySelector('input').getAttribute('id'),
+                'install-button-foo', 'id is set');
+    assert.equal(root.querySelector('label').getAttribute('for'),
+                'install-button-foo', 'for attribute matches id');
+  });
+
+  it('should throw on bogus status', () => {
+    assert.throws(() => {
+      renderButton({status: 'BOGUS'});
+    }, Error, 'Invalid add-on status');
+  });
+
+  it('should call installTheme function on click when uninstalled theme', () => {
+    const installTheme = sinon.spy();
+    const guid = 'test-guid';
+    const name = 'hai';
+    const button = renderButton({installTheme, type: THEME_TYPE, guid, name, status: UNINSTALLED});
+    const themeData = button.refs.themeData;
+    const root = findDOMNode(button);
+    Simulate.click(root);
+    assert(installTheme.calledWith(themeData, guid, name));
+  });
+
+  it('should call install function on click when uninstalled', () => {
+    const guid = '@foo';
+    const name = 'hai';
+    const install = sinon.spy();
+    const installURL = 'https://my.url/download';
+    const button = renderButton({guid, install, installURL, name, status: UNINSTALLED});
+    const root = findDOMNode(button);
+    Simulate.click(root);
+    assert(install.calledWith({guid, installURL, name}));
+  });
+
+  it('should call uninstall function on click when installed', () => {
+    const guid = '@foo';
+    const installURL = 'https://my.url/download';
+    const name = 'hai';
+    const type = 'whatevs';
+    const uninstall = sinon.spy();
+    const button = renderButton({guid, installURL, name, status: INSTALLED, type, uninstall});
+    const root = findDOMNode(button);
+    Simulate.click(root);
+    assert(uninstall.calledWith({guid, installURL, name, type}));
+  });
+});
